Replace Spliterator implementation with AbstractSpliterator
