+++ src/test/java/org/jayield/QueryIterateTest.java
@@ -44,26 +44,6 @@ import static org.testng.Assert.assertTrue;
 public class QueryIterateTest {
 
     @Test
-    public void testFromStream() {
-        Integer[] src = {1, 2, 3, 4, 5, 6, 7, 8, 9};
-        Iterator<Integer> expected = Stream.of(src).iterator();
-        Query<Integer> nrs = fromStream(Stream.of(1, 2, 3, 4, 5, 6, 7, 8, 9));
-        while(nrs.hasNext()) {
-            assertEquals(nrs.next(), expected.next());
-        }
-        assertFalse(expected.hasNext());
-    }
-    @Test
-    public void testFromAndToStream() {
-        Integer[] src = {1, 2, 3, 4, 5, 6, 7, 8, 9};
-        Iterator<Integer> expected = Stream.of(src).iterator();
-        Query<Integer> nrs = fromStream(Stream.of(1, 2, 3, 4, 5, 6, 7, 8, 9));
-        Spliterator<Integer> actual = nrs.toStream().spliterator();
-        while(actual.tryAdvance(curr -> assertEquals(curr, expected.next()))) {}
-        assertFalse(expected.hasNext());
-    }
-
-    @Test
     public void testZip() {
         Object[] expected = {"a1", "b2", "c3", "d4", "e5", "f6", "g7"};
         Query<Integer> nrs = Query.of(1, 2, 3, 4, 5, 6, 7, 8, 9);
